backend/requirements.txt
flask
flask-cors
werkzeug
PyPDF2
google-generativeai
python-dotenv
pymongo[srv]
certifi
pdf2image
pytesseract
